# combined_swagger.yaml
openapi: 3.0.3
info:
  title: TAAFT Backend API
  description: API for the TAAFT (Tools Analysis and Functionality Testing) application
  version: 1.0.0
  contact:
    name: API Support
    email: support@taaft.com
servers:
- url: http://localhost:8000
  description: Local development server
- url: https://api.taaft.com
  description: Production server
...
paths:
...
  /api/glossary/terms/{term_id}:
    get:
      tags:
      - glossary
      summary: Get a glossary term
      description: 'Get a specific glossary term by ID.

        No authentication required (free tier access).

        '
      parameters:
      - name: term_id
        in: path
        description: ID of the glossary term
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Glossary term
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlossaryTermResponse'
        '404':
          description: Glossary term not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - glossary
      summary: Update a glossary term
      description: 'Update an existing glossary term.

        No authentication required (free tier access).

        '
      parameters:
      - name: term_id
        in: path
        description: ID of the glossary term
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlossaryTermUpdate'
      responses:
        '200':
          description: Updated glossary term
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlossaryTermResponse'
        '404':
          description: Glossary term not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Glossary term with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - glossary
      summary: Delete a glossary term
      description: 'Delete a glossary term.

        No authentication required (free tier access).

        '
      parameters:
      - name: term_id
        in: path
        description: ID of the glossary term
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Glossary term deleted
        '404':
          description: Glossary term not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/glossary/categories:
    get:
      tags:
      - glossary
      summary: Get all categories
      description: 'Get all unique categories from the glossary terms.

        No authentication required (free tier access).

        '
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/categories/{category_id}:
    get:
      tags:
      - Categories
      summary: Get category by ID
      description: Get a category by its ID
      operationId: getCategoryById
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: string
        description: ID of the category to fetch
      responses:
        '200':
          description: Category object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found
  /api/categories/slug/{slug}:
    get:
      tags:
      - Categories
      summary: Get category by slug
      description: Get a category by its slug
      operationId: getCategoryBySlug
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        description: Slug of the category to fetch
      responses:
        '200':
          description: Category object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found
  /api/terms/define:
    post:
      tags:
      - Terms
      summary: Define a term
      description: Get a definition and examples for a specific term or concept
      operationId: defineTerm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TermDefinitionRequest'
      responses:
        '200':
          description: Term definition successfully generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermDefinitionResponse'
        '400':
          description: Invalid request
        '500':
          description: Error getting response from language model
  /api/terms/history:
    get:
      tags:
      - Terms
      summary: Get term definition history
      description: Get history of previously requested term definitions
      operationId: getTermHistory
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          type: string
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        '200':
          description: List of term definition history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TermDefinition'
  /api/terms/popular:
    get:
      tags:
      - Terms
      summary: Get popular terms
      description: Get most frequently requested terms
      operationId: getPopularTerms
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
      responses:
        '200':
          description: List of popular terms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PopularTerm'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GlossaryTermResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the glossary term
        name:
          type: string
          description: The name/title of the glossary term
        definition:
          type: string
          description: Detailed definition of the term
        related_terms:
          type: array
          items:
            type: string
          description: List of related term names
        tool_references:
          type: array
          items:
            type: string
          description: List of tool IDs that are related to this term
        categories:
          type: array
          items:
            type: string
          description: Categories this term belongs to
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
      - id
      - name
      - definition
      - created_at
      - updated_at
...
    GlossaryTermUpdate:
      type: object
      properties:
        name:
          type: string
          description: The name/title of the glossary term
        definition:
          type: string
          description: Detailed definition of the term
        related_terms:
          type: array
          items:
            type: string
          description: List of related term names
        tool_references:
          type: array
          items:
            type: string
          description: List of tool IDs that are related to this term
        categories:
          type: array
          items:
            type: string
          description: Categories this term belongs to
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          description: Category ID
        name:
          type: string
          description: Category name
        slug:
          type: string
          description: URL-friendly version of the category name
        count:
          type: integer
          description: Number of tools in this category
      required:
      - id
      - name
      - slug
    TermDefinitionRequest:
      type: object
      required:
      - term
      properties:
        term:
          type: string
          description: The term or concept to define
        user_id:
          type: string
          description: Optional user ID for tracking history
        model:
          $ref: '#/components/schemas/TermModelType'
    TermDefinitionResponse:
      type: object
      properties:
        term:
          type: string
          description: The term that was defined
        description:
          type: string
          description: Brief description of the term
        examples:
          type: array
          items:
            type: string
          description: Examples demonstrating the term
        id:
          type: string
          description: Unique ID for this definition
        timestamp:
          type: string
          format: date-time
          description: When this definition was generated
        model:
          $ref: '#/components/schemas/TermModelType'
    TermDefinition:
      type: object
      properties:
        id:
          type: string
          description: Unique ID for this definition
        term:
          type: string
          description: The term that was defined
        description:
          type: string
          description: Brief description of the term
        examples:
          type: array
          items:
            type: string
          description: Examples demonstrating the term
        user_id:
          type: string
          description: User who requested this definition
        timestamp:
          type: string
          format: date-time
          description: When this definition was generated
        model:
          $ref: '#/components/schemas/TermModelType'
    PopularTerm:
      type: object
      properties:
        term:
          type: string
          description: The popular term
        count:
          type: integer
          description: Number of times this term has been requested
        last_requested:
          type: string
          format: date-time
          description: When this term was last requested
    TermModelType:
      type: string
      enum:
      - gpt4
      - claude
      - llama
      - default
      default: default
security:
- bearerAuth: []